requirements.txt
accelerate
aioboto3
aiohttp

# database packages
graspologic
hnswlib
nano-vectordb
neo4j
networkx
ollama
openai
oracledb
pymilvus
pymongo
pymysql
pyvis
# lmdeploy[all]
sqlalchemy
tenacity


# LLM packages
tiktoken
torch
transformers
xxhash

# New Dependencies
google-generativeai
pillow
sentence-transformers
python-dotenv

# Web Framework & API
fastapi
uvicorn
python-multipart

# Document Processing
pypdf
python-docx

# Vector Database
chromadb
pydantic
tqdm
